Update theme using Dark Purple as a basis
Fix a bunch of color issues, including being too flat of a color scheme.
Add inverted colors and lighter alternatives to make things easier on
the eyes.

This change is being made because the color scheme was way too flat. It
was all one color essentially and any differentiation was very hard to
see.